requirements.txt
tomlkit
pathlib
reporoulette
pygithub
xml
stat
pandas
matplotlib
google
openai
anthropic